worked on users

+++ src/ois/model/impl/ModelManagerImpl.java
@@ -15,6 +15,9 @@ import ois.model.ImageFile;
 import ois.model.ModelManager;
 
 import com.google.appengine.api.datastore.Key;
+import com.google.appengine.api.users.User;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 
 public class ModelManagerImpl implements ModelManager {
 	private static final Logger log = Logger.getLogger(ModelManagerImpl.class.getName());
@@ -23,6 +26,15 @@ public class ModelManagerImpl implements ModelManager {
 	 * @see ois.model.impl.ModelManager#saveImage(ois.model.ImageFile)
 	 */
 	public void saveImageFile(ImageFile imageFile, PersistenceManager pm) throws PersistanceManagerException{
+		UserService userService = UserServiceFactory.getUserService();
+		//create album
+		if(imageFile.getKey() == null){
+			//admin creating an image
+			imageFile.setGlobal(userService.isUserAdmin());
+		}else if(imageFile.isGlobal() && !userService.isUserAdmin())
+			//trying to change admin object with normal user. return
+			return;
+
 		try {
             pm.makePersistent(imageFile);
         }catch(Exception e){
@@ -38,6 +50,15 @@ public class ModelManagerImpl implements ModelManager {
 	 * @see ois.model.impl.ModelManager#saveImage(ois.model.ImageFile)
 	 */
 	public void saveImageData(ImageData imageData, PersistenceManager pm) throws ImageDataTooBigException{
+		UserService userService = UserServiceFactory.getUserService();
+		//create album
+		if(imageData.getKey() == null){
+			//admin creating an album
+			imageData.setGlobal(userService.isUserAdmin());
+		}else if(imageData.isGlobal() && !userService.isUserAdmin())
+			//trying to change admin object with normal user. return
+			return;
+
 		try {
             pm.makePersistent(imageData);
 		}catch(Exception e){
@@ -52,15 +73,38 @@ public class ModelManagerImpl implements ModelManager {
 	/* (non-Javadoc)
 	 * @see ois.model.impl.ModelManager#getAlbums()
 	 */
+	@SuppressWarnings("unchecked")
 	public Iterable<AlbumFile> getAlbums(PersistenceManager pm){
-		Iterable<AlbumFile> albums = pm.getExtent(AlbumFile.class);
-		return albums;
+		//Iterable<AlbumFile> albums = pm.getExtent(AlbumFile.class);
+		//return albums;
+		UserService userService = UserServiceFactory.getUserService();
+		User user = userService.getCurrentUser();
+		
+		Query query = pm.newQuery(AlbumFile.class);
+		if(!userService.isUserAdmin()){
+			query.setFilter("owner == ownerParam || owner == null");
+			query.declareParameters("com.google.appengine.api.users.User ownerParam, int isAdminParam");
+		}
+		List<AlbumFile> albumFileList = (List<AlbumFile>) query.execute(user,userService.isUserAdmin()?1:0);
+		return albumFileList;
 	}
 	
 	/* (non-Javadoc)
 	 * @see ois.model.impl.ModelManager#saveAlbum(ois.model.AlbumFile)
 	 */
 	public void saveAlbum(AlbumFile album,PersistenceManager pm) throws PersistanceManagerException{
+		UserService userService = UserServiceFactory.getUserService();
+		User user = userService.getCurrentUser();
+		//create album
+		if(album.getKey() == null){
+			//admin creating an album
+			if(userService.isUserAdmin())
+				user = null;
+			album.setOwner(user);
+		}else if(album.getOwner() == null && !userService.isUserAdmin()){
+			//trying to change admin object with normal user. return
+			pm.detachCopy(album);
+			return;}
 		try {
 			pm.makePersistent(album);
         }catch(Exception e){
@@ -97,7 +141,11 @@ public class ModelManagerImpl implements ModelManager {
 	 */
 	public void deleteAlbumFile(Key key,PersistenceManager pm) throws PersistanceManagerException {
         AlbumFile albumFile = getAlbumFile(key, pm);
-		try {
+		
+		if(albumFile.getOwner()==null && !UserServiceFactory.getUserService().isUserAdmin())
+			return;
+
+        try {
         	pm.deletePersistent(albumFile);
         }catch(Exception e){
         	PersistanceManagerException pme = new PersistanceManagerException("Album [" + albumFile.getName() + "] could not be deleted",e);
@@ -210,7 +258,11 @@ public class ModelManagerImpl implements ModelManager {
 	}
 	@Override
 	public void deleteImageData(ImageData imageData, PersistenceManager pm) throws PersistanceManagerException {
-        try {
+        
+		if(imageData.isGlobal() && !UserServiceFactory.getUserService().isUserAdmin())
+			return;
+
+		try {
         	pm.deletePersistent(imageData);
         }catch(Exception e){
         	PersistanceManagerException pme = new PersistanceManagerException("Image data [" + imageData.getKey() + "] could not be deleted",e);
@@ -225,6 +277,10 @@ public class ModelManagerImpl implements ModelManager {
 	}
 	@Override
 	public void deleteImageFile(ImageFile imageFile, PersistenceManager pm) throws PersistanceManagerException {
+		
+		if(imageFile.isGlobal() && !UserServiceFactory.getUserService().isUserAdmin())
+			return;
+		
 		try {
         	pm.deletePersistent(imageFile);
         }catch(Exception e){
